lots of stuff

+++ db/migrate/20230214200108_create_trails.rb
@@ -3,7 +3,7 @@ class CreateTrails < ActiveRecord::Migration[6.1]
     create_table :trails do |t|
       t.string :name
       t.integer :elevation_gain
-      t.integer :length
+      t.float :length
 
       t.timestamps
     end
